@@ -3,22 +3,23 @@ import 'package:flutter/material.dart';
 
 class CloudStorageInfo {
   final String svgSrc, title, totalStorage;
-  final int numOfFiels, percentage;
+  final int numOfFiles, percentage;
   final Color color;
 
-  CloudStorageInfo(
-      {this.svgSrc,
-      this.title,
-      this.totalStorage,
-      this.numOfFiels,
-      this.percentage,
-      this.color});
+  CloudStorageInfo({
+    required this.svgSrc,
+    required this.title,
+    required this.totalStorage,
+    required this.numOfFiles,
+    required this.percentage,
+    required this.color,
+  });
 }
 
-List demoMyFiels = [
+List demoMyFiles = [
   CloudStorageInfo(
     title: "Documents",
-    numOfFiels: 1328,
+    numOfFiles: 1328,
     svgSrc: "assets/icons/Documents.svg",
     totalStorage: "1.9GB",
     color: primaryColor,
@@ -26,7 +27,7 @@ List demoMyFiels = [
   ),
   CloudStorageInfo(
     title: "Google Drive",
-    numOfFiels: 1328,
+    numOfFiles: 1328,
     svgSrc: "assets/icons/google_drive.svg",
     totalStorage: "2.9GB",
     color: Color(0xFFFFA113),
@@ -34,7 +35,7 @@ List demoMyFiels = [
   ),
   CloudStorageInfo(
     title: "One Drive",
-    numOfFiels: 1328,
+    numOfFiles: 1328,
     svgSrc: "assets/icons/one_drive.svg",
     totalStorage: "1GB",
     color: Color(0xFFA4CDFF),
@@ -42,7 +43,7 @@ List demoMyFiels = [
   ),
   CloudStorageInfo(
     title: "Documents",
-    numOfFiels: 5328,
+    numOfFiles: 5328,
     svgSrc: "assets/icons/drop_box.svg",
     totalStorage: "7.3GB",
     color: Color(0xFF007EE5),
