@@ -1,7 +1,12 @@
 class RecentFile {
   final String icon, title, date, size;
 
-  RecentFile({this.icon, this.title, this.date, this.size});
+  RecentFile({
+    required this.icon,
+    required this.title,
+    required this.date,
+    required this.size,
+  });
 }
 
 List demoRecentFiles = [
@@ -19,7 +24,7 @@ List demoRecentFiles = [
   ),
   RecentFile(
     icon: "assets/icons/doc_file.svg",
-    title: "Documetns",
+    title: "Documents",
     date: "23-02-2021",
     size: "32.5mb",
   ),
@@ -37,12 +42,12 @@ List demoRecentFiles = [
   ),
   RecentFile(
     icon: "assets/icons/pdf_file.svg",
-    title: "Sals PDF",
+    title: "Sales PDF",
     date: "25-02-2021",
     size: "3.5mb",
   ),
   RecentFile(
-    icon: "assets/icons/excle_file.svg",
+    icon: "assets/icons/excel_file.svg",
     title: "Excel File",
     date: "25-02-2021",
     size: "34.5mb",
