@@ -1,8 +1,33 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import 'package:admin/controllers/MenuController.dart' as menuController;
+import 'package:admin/responsive.dart';
+
+import '../dashboard/dashboard_screen.dart';
+import 'components/side_menu.dart';
 
 class MainScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Scaffold();
+    return Scaffold(
+      key: context.read<menuController.MenuController>().scaffoldKey,
+      drawer: SideMenu(),
+      body: SafeArea(
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            if (Responsive.isDesktop(context))
+              Expanded(
+                child: SideMenu(),
+              ),
+            Expanded(
+              flex: 5,
+              child: DashboardScreen(),
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }
